refactor: extract lexer+minify into internal/lua package
