+++ internal/lua/lexer.go
@@ -1,4 +1,4 @@
-package bundler
+package lua
 
 import "strings"
