+++ internal/lua/lexer_test.go
@@ -1,4 +1,4 @@
-package bundler
+package lua
 
 import (
 	"testing"
